+++ buildpacks/ruby/src/layers.rs
@@ -1,5 +1,6 @@
 mod bundle_download_layer;
 mod bundle_install_layer;
+pub(crate) mod metrics_agent_install;
 mod ruby_install_layer;
 
 pub(crate) use self::bundle_download_layer::BundleDownloadLayer;