+++ buildpacks/ruby/tests/fixtures/barnes_app/config/puma.rb
@@ -1,0 +1,9 @@
+require 'barnes'
+
+before_fork do
+  # worker specific setup
+
+  Barnes.start # Must have enabled worker mode for this to block to be called
+end
+
+workers 2